+++ docs/src/markdown/about/changelog.md
@@ -9,6 +9,10 @@
 - **NEW**: Arithmatex: Add options to override HTML element container type of inline and block math.
 - **NEW**: Arithmatex: Add new formatter functions intended to replace old math fenced/inline block formatters. New
   formatter functions are configurable. All others are marked as deprecated and will be removed at some future date.
+- **NEW**: Highlight: Add support for Pygments option `linespans`.
+- **NEW**: Highlight: Add support for Pygments option `filename`. Also add new option `auto_filename` which will use
+  the current lexer's name as the title. Also add `auto_filename_mapping` which allows a user to override a given
+  lexer's name automatically.
 - **FIX**: BetterEm: Fix some complex cases related to "smart" logic. (#1413)
 - **FIX**: Highlight: Fix issue that occurs when showing only nth line numbers and using `pymdownx-inline`. Lines not
   showing a line number would not render with the proper leading space.